requirements.txt
discord.py
python-dotenv
aiohttp
motor
colorlog
emoji